Adds: backward compatibility of root file to RDF for reprocessing

+++ DreamAN/Cuts/EventCut.cxx
@@ -71,8 +71,8 @@ EventCut* EventCut::PhotonCuts() {
 
 // Filtering logic with Fiducial cuts
 bool EventCut::operator()(const std::vector<int>& pid, const std::vector<float>& px, const std::vector<float>& py, const std::vector<float>& pz, const std::vector<float>& vx,
-                          const std::vector<float>& vy, const std::vector<float>& vz, const std::vector<float>& vt, const std::vector<int>& charge, const std::vector<float>& beta,
-                          const std::vector<float>& chi2pid, const std::vector<int>& status, const std::vector<int>& REC_Track_pass_fid) const {
+                          const std::vector<float>& vy, const std::vector<float>& vz, const std::vector<float>& vt, const std::vector<short>& charge, const std::vector<float>& beta,
+                          const std::vector<float>& chi2pid, const std::vector<short>& status, const std::vector<int>& REC_Track_pass_fid) const {
   int pidCount = 0;
   for (size_t i = 0; i < pid.size(); ++i) {
     // Skip trivial momentum
@@ -82,7 +82,7 @@ bool EventCut::operator()(const std::vector<int>& pid, const std::vector<float>&
 
     if (p2 < 1e-4f) continue; // Equivalent to momentum < 0.01
 
-    if (pid[i] != fSelectedPID || static_cast<int8_t>(charge[i]) != fCharge || REC_Track_pass_fid[i] != 1 || !IsInRange(chi2pid[i], fMinChi2PID, fMaxChi2PID))
+    if (pid[i] != fSelectedPID || charge[i] != fCharge || REC_Track_pass_fid[i] != 1 || !IsInRange(chi2pid[i], fMinChi2PID, fMaxChi2PID))
       continue;
 
     const float momentum = std::sqrt(p2);
